requirements.txt
# Core Agent Lightning framework
agentlightning>=0.1.0

# Google Gemini SDK
google-generativeai>=0.21.0

# Vector Store
qdrant-client>=1.7.0

# Graph Store
neo4j>=5.15.0

# Cache
redis>=5.0.0

# Utilities
cachetools>=5.3.0
aiohttp>=3.9.0
asyncio-throttle>=1.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Existing Crawl4AI dependencies
crawl4ai>=0.1.0
playwright>=1.40.0
selenium>=4.16.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Data processing
pandas>=2.0.0

# Environment
python-dotenv>=1.0.0
